backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
pillow
imagehash
numpy
google-generativeai
python-dotenv
aiofiles
